+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - enhanced MCP server configuration with possibility to update the whole config
 - added support for Ollama provider in Agent
 - replace system prompt with custom instructions for Agent
+- fixed issue with `npx` in MCP servers on Windows
 
 ## [0.8.0]